chore: moves parser to its own package
